Split and simplify ragged_conv1d_jax and ragged_gated_delta_rule_jax
